Dates: make the package-period `Timestamp` test 32-bit clean
The year `10^18` is `Int32` arithmetic on 32-bit hosts and wraps to
-1486618624, a year that `Date` can represent, so `Date(T(10^18))` did not
throw. Use `Int64(10)^18`. Failed in CI on i686 Windows:
https://buildkite.com/julialang/julia-pr/builds/2861#01a0d5b6-d424-4a7f-9de1-47ec5f4f5f55

+++ stdlib/Dates/test/timestamps.jl
@@ -633,8 +633,8 @@ Dates.tons(p::TestPicosecond) = Dates.value(p) // big(1000)
     @test Date(before) == Date(1969, 12, 31)
     @test Dates.nanosecond(before) == 999
     @test floor(before, Nanosecond) == convert(T, Nanosecond(-1))
-    @test year(T(10^18)) == 10^18
-    @test_throws InexactError Date(T(10^18))
+    @test year(T(Int64(10)^18)) == Int64(10)^18
+    @test_throws InexactError Date(T(Int64(10)^18))
     @test_throws ArgumentError T(Year(2026), TestPicosecond(1))
     @test_throws ArgumentError T(TestPicosecond(1))
     @test isempty(Test.detect_ambiguities(TimestampPeriodExtensionTests; recursive=true))
